Major code change for the collector generic code. Remove old exporter flowsource code. Add new exporter/flowsource code to improve handling of multiple flow sources with lots of exporters, which are now organised in hash tables. Add --enable-lto switch to improve code linking Add --enable-native in order to generate locally more efficient code.

+++ src/nfpcapd/packet_bpf.c
@@ -103,7 +103,7 @@ static void CloseSocket(packetParam_t *param) {
 }  // End of CloseSocket
 
 // live device
-int setup_bpf_live(packetParam_t *param, char *device, char *filter, int snaplen, int buffsize, int to_ms) {
+int setup_bpf_live(packetParam_t *param, char *device, char *filter, unsigned snaplen, size_t buffsize, int to_ms) {
     param->pcap_dev = NULL;
     param->bpf = 0;
     int bpf = open_bpf();
