+++ src/nfpcapd/packet_linux.c
@@ -145,7 +145,7 @@ static int InitRing(packetParam_t *param, char *device) {
 
 }  // End of InitRing
 
-int setup_linux_live(packetParam_t *param, char *device, char *filter, int snaplen, int buffsize, int to_ms) {
+int setup_linux_live(packetParam_t *param, char *device, char *filter, unsigned snaplen, size_t buffsize, int to_ms) {
     param->pcap_dev = NULL;
     param->fd = 0;
